pdf funciton moved into new script and names changed
