+++ tsfresh/feature_extraction/__init__.py
@@ -2,6 +2,6 @@
 The :mod:`tsfresh.feature_extraction` module contains methods to extract the features from the time series
 """
 
-from tsfresh.feature_extraction.extraction import extract_features, extract_features_on_sub_features
+from tsfresh.feature_extraction.extraction import extract_features, extract_feature_dynamics
 from tsfresh.feature_extraction.settings import ComprehensiveFCParameters, MinimalFCParameters, \
     EfficientFCParameters
